refactor: modernize codebase with Go 1.22+ idioms and improve code quality
- Use for-range over integer iteration (for i := 0; i < n; i++ → for range n)
- Replace string index/slice patterns with strings.Cut, CutPrefix, CutSuffix
- Use maps.Copy for map merging and errors.As for error type checking
- Use min() builtin and strings.SplitSeq for sequential iteration
- Improve error handling with proper wrapping (%w) and add missing checks
- Extract common table mutation logic into shared mutateTableElements function
- Fix code style issues: line length, naming consistency, formatting
- Add nolint directives for intentional global variables
- Use b.Loop() for benchmarks (Go 1.24+)

+++ internal/pptxxml/comments_xml.go
@@ -15,7 +15,7 @@ func CommentAuthorsXML(authors []comments.Author) string {
 	for _, author := range authors {
 		b.WriteString(`
 <p:cmAuthor id="`)
-		b.WriteString(strconv.FormatInt(int64(author.ID), 10))
+		b.WriteString(strconv.FormatInt(author.ID, 10))
 		b.WriteString(`" name="`)
 		b.WriteString(Escape(author.Name))
 		b.WriteString(`" initials="`)
@@ -41,16 +41,16 @@ func CommentsXML(slideComments []comments.Comment) string {
 		dt := cm.Date.Format("2006-01-02T15:04:05.000") // ISO 8601
 		b.WriteString(`
 <p:cm authorId="`)
-		b.WriteString(strconv.FormatInt(int64(cm.AuthorID), 10))
+		b.WriteString(strconv.FormatInt(cm.AuthorID, 10))
 		b.WriteString(`" dt="`)
 		b.WriteString(dt)
 		b.WriteString(`" idx="`)
 		b.WriteString(strconv.FormatInt(int64(cm.Index), 10))
 		b.WriteString(`">
 <p:pos x="`)
-		b.WriteString(strconv.FormatInt(int64(cm.X), 10))
+		b.WriteString(strconv.FormatInt(cm.X, 10))
 		b.WriteString(`" y="`)
-		b.WriteString(strconv.FormatInt(int64(cm.Y), 10))
+		b.WriteString(strconv.FormatInt(cm.Y, 10))
 		b.WriteString(`"/>
 <p:text>`)
 		b.WriteString(Escape(cm.Text))
